// src/mocks/productList/type.ts
export interface productListType {
  createdDate: number;
  updatedDate: number;
  productIndex: number;
  productNumber: number;
  barcodeNumber: string;
  productName: string;
  productOption: string;
  packageUnit: string;
  packageMinUnitQuan: number;
  minimunUnit: string;
  processor: string;
}